addressed by comments
