+++ core/src/main/scala/kafka/server/ForwardingManager.scala
@@ -25,7 +25,8 @@ import org.apache.kafka.common.errors.TimeoutException
 import org.apache.kafka.common.metrics.Metrics
 import org.apache.kafka.common.protocol.Errors
 import org.apache.kafka.common.requests.{AbstractRequest, AbstractResponse, EnvelopeRequest, EnvelopeResponse, RequestContext, RequestHeader}
-import org.apache.kafka.server.{ControllerRequestCompletionHandler, NodeToControllerChannelManager}
+import org.apache.kafka.server.common.ControllerRequestCompletionHandler
+import org.apache.kafka.server.common.serialization.NodeToControllerChannelManager
 
 import java.util.concurrent.TimeUnit
 import scala.jdk.OptionConverters.RichOptional
