+++ core/src/test/scala/unit/kafka/coordinator/transaction/ProducerIdManagerTest.scala
@@ -18,8 +18,8 @@ package kafka.coordinator.transaction
 
 import kafka.zk.{KafkaZkClient, ProducerIdBlockZNode}
 import org.apache.kafka.common.KafkaException
-import org.apache.kafka.server.NodeToControllerChannelManager
 import org.apache.kafka.server.common.ProducerIdsBlock
+import org.apache.kafka.server.common.serialization.NodeToControllerChannelManager
 import org.junit.jupiter.api.Assertions._
 import org.junit.jupiter.api.Test
 import org.mockito.ArgumentCaptor
