+++ core/src/test/scala/unit/kafka/server/AutoTopicCreationManagerTest.scala
@@ -40,7 +40,8 @@ import org.apache.kafka.coordinator.group.{GroupCoordinator, GroupCoordinatorCon
 import org.apache.kafka.coordinator.share.ShareCoordinator
 import org.apache.kafka.server.config.{ServerConfigs, ShareCoordinatorConfig}
 import org.apache.kafka.coordinator.transaction.TransactionLogConfig
-import org.apache.kafka.server.{ControllerRequestCompletionHandler, NodeToControllerChannelManager}
+import org.apache.kafka.server.common.ControllerRequestCompletionHandler
+import org.apache.kafka.server.common.serialization.NodeToControllerChannelManager
 import org.junit.jupiter.api.Assertions.{assertEquals, assertThrows, assertTrue}
 import org.junit.jupiter.api.{BeforeEach, Test}
 import org.mockito.ArgumentMatchers.any
