+++ core/src/test/scala/unit/kafka/server/MockNodeToControllerChannelManager.scala
@@ -19,7 +19,8 @@ package kafka.server
 import org.apache.kafka.clients.{ClientResponse, MockClient, NodeApiVersions}
 import org.apache.kafka.common.protocol.Errors
 import org.apache.kafka.common.requests.AbstractRequest
-import org.apache.kafka.server.{ControllerRequestCompletionHandler, NodeToControllerChannelManager}
+import org.apache.kafka.server.common.ControllerRequestCompletionHandler
+import org.apache.kafka.server.common.serialization.NodeToControllerChannelManager
 import org.apache.kafka.server.util.MockTime
 
 import java.util.Optional
