+++ server/src/main/java/org/apache/kafka/server/AssignmentsManager.java
@@ -34,7 +34,9 @@ import org.apache.kafka.image.MetadataImage;
 import org.apache.kafka.image.TopicImage;
 import org.apache.kafka.queue.EventQueue;
 import org.apache.kafka.queue.KafkaEventQueue;
+import org.apache.kafka.server.common.ControllerRequestCompletionHandler;
 import org.apache.kafka.server.common.TopicIdPartition;
+import org.apache.kafka.server.common.serialization.NodeToControllerChannelManager;
 import org.apache.kafka.server.metrics.KafkaYammerMetrics;
 
 import com.yammer.metrics.core.Gauge;
